Cast releaseId and projectId to long

+++ src/main/scala/npmvuln/jobs/ProjectDfBuilder.scala
@@ -3,6 +3,7 @@ package npmvuln.jobs
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.functions.row_number
 import org.apache.spark.sql.expressions.Window
+import org.apache.spark.sql.types._
 
 object ProjectDfBuilder {
 
@@ -16,6 +17,7 @@ object ProjectDfBuilder {
       .distinct
 
       // Add ID field for every package (0 upwards)
-      .withColumn("ProjectId", row_number.over(Window.orderBy("Project")) - 1)
+      .withColumn("ProjectId",
+        row_number.over(Window.orderBy("Project")).cast(LongType) - 1)
   }
 }
